+++ graspit-reporters/packages/graspit-dashboard/src/components/utils/Badge.tsx
@@ -2,16 +2,16 @@ import type { BadgeProps } from 'antd/lib/badge/index';
 import Badge from 'antd/lib/badge/index';
 import React, { type ReactNode } from 'react';
 
-export default function BadgeForSuiteType(props: {
+export default function BadgeForSuiteType(properties: {
     text: BadgeProps['count'];
     color: BadgeProps['color'];
     size?: BadgeProps['size'];
 }): ReactNode {
     return (
         <Badge
-            color={props.color}
-            count={props.text}
-            size={props.size}
+            color={properties.color}
+            count={properties.text}
+            size={properties.size}
             style={{ fontWeight: 'bold', color: 'white' }}
         />
     );