+++ graspit-reporters/packages/graspit-dashboard/src/components/utils/Datetime/relative-time.tsx
@@ -20,21 +20,21 @@ dayjs.extend(relativeTime);
 dayjs.extend(duration);
 dayjs.extend(advancedFormat);
 
-export default function RelativeTo(props: {
+export default function RelativeTo(properties: {
     dateTime: Dayjs;
     wrt?: Dayjs;
     secondDateTime?: Dayjs;
     format?: string;
     style?: CSSProperties;
 }): ReactNode {
-    const [emblaRef, emblaApi] = useEmblaCarousel({ loop: true }, [
+    const [emblaReference, emblaApi] = useEmblaCarousel({ loop: true }, [
         Autoplay({ stopOnInteraction: false }),
     ]);
     const [hover, setHover] = useState<boolean>(false);
     const formatter = (): string =>
-        props.wrt != null
-            ? props.dateTime.from(props.wrt)
-            : props.dateTime.fromNow();
+        properties.wrt == undefined
+            ? properties.dateTime.fromNow()
+            : properties.dateTime.from(properties.wrt);
 
     const [formatted, setFormatted] = useState(formatter());
     const [isClient, setIsClient] = useState(false);
@@ -51,35 +51,26 @@ export default function RelativeTo(props: {
     return (
         <div
             className={carouselStyles.embla}
-            ref={emblaRef}
+            ref={emblaReference}
             style={{
                 maxWidth: '120px',
                 cursor: hover ? 'grabbing' : 'grab',
-                ...(props.style ?? {}),
+                ...properties.style,
             }}
         >
             <div suppressHydrationWarning className={carouselStyles.container}>
                 <Typography className={carouselStyles.slide}>
-                    {`${props.dateTime.format(
-                        props.format ?? timeFormatUsed,
+                    {`${properties.dateTime.format(
+                        properties.format ?? timeFormatUsed,
                     )} ${
-                        props.secondDateTime != null
-                            ? ` - ${props.secondDateTime.format(
-                                  props.format ?? timeFormatUsed,
+                        properties.secondDateTime == undefined
+                            ? ''
+                            : ` - ${properties.secondDateTime.format(
+                                  properties.format ?? timeFormatUsed,
                               )}`
-                            : ''
                     }`}
                 </Typography>
-                {props.wrt != null ? (
-                    <Tooltip
-                        title={`Relative to ${props.wrt.format(
-                            props.format ?? timeFormatUsed,
-                        )}`}
-                        className={carouselStyles.slide}
-                    >
-                        {isClient ? formatted : ''}
-                    </Tooltip>
-                ) : (
+                {properties.wrt == undefined ? (
                     <span
                         className={carouselStyles.slide}
                         onMouseEnter={() => {
@@ -88,17 +79,26 @@ export default function RelativeTo(props: {
                     >
                         {isClient ? formatted : ''}
                     </span>
+                ) : (
+                    <Tooltip
+                        title={`Relative to ${properties.wrt.format(
+                            properties.format ?? timeFormatUsed,
+                        )}`}
+                        className={carouselStyles.slide}
+                    >
+                        {isClient ? formatted : ''}
+                    </Tooltip>
                 )}
             </div>
         </div>
     );
 }
 
-export function HumanizeDuration(props: {
+export function HumanizeDuration(properties: {
     duration?: Duration;
     style?: CSSProperties;
 }): ReactNode {
-    const [emblaRef, emblaApi] = useEmblaCarousel({ loop: true }, [
+    const [emblaReference, emblaApi] = useEmblaCarousel({ loop: true }, [
         Autoplay({ stopOnInteraction: false }),
     ]);
     const [hover, setHover] = useState<boolean>(false);
@@ -114,12 +114,12 @@ export function HumanizeDuration(props: {
     return (
         <div
             className={carouselStyles.embla}
-            ref={emblaRef}
+            ref={emblaReference}
             style={{
                 maxWidth: '150px',
                 minWidth: '100px',
                 cursor: hover ? 'grabbing' : 'grab',
-                ...(props.style ?? {}),
+                ...properties.style,
             }}
         >
             <div suppressHydrationWarning className={carouselStyles.container}>
@@ -127,13 +127,15 @@ export function HumanizeDuration(props: {
                     suppressHydrationWarning
                     className={carouselStyles.slide}
                 >
-                    {`${props?.duration?.asSeconds().toFixed(2) ?? '--'} s`}
+                    {`${
+                        properties?.duration?.asSeconds().toFixed(2) ?? '--'
+                    } s`}
                 </Typography>
                 <Typography
                     suppressHydrationWarning
                     className={carouselStyles.slide}
                 >
-                    {props?.duration?.humanize() ?? '--'}
+                    {properties?.duration?.humanize() ?? '--'}
                 </Typography>
             </div>
         </div>