+++ graspit-reporters/packages/graspit-dashboard/src/components/utils/renderers.tsx
@@ -1,49 +1,54 @@
 import { type Dayjs } from 'dayjs';
 import React, { type CSSProperties, type ReactNode } from 'react';
 import { type Duration } from 'dayjs/plugin/duration';
-import type { statusOfEntity } from 'src/types/sessionRecords';
+import type { statusOfEntity } from 'src/types/session-records';
 import CheckCircleFilled from '@ant-design/icons/CheckCircleFilled';
 import ReloadOutlined from '@ant-design/icons/ReloadOutlined';
 import CloseOutlined from '@ant-design/icons/CloseOutlined';
 import WarningFilled from '@ant-design/icons/WarningFilled';
 import LoadingOutlined from '@ant-design/icons/LoadingOutlined';
-import type { possibleEntityNames } from 'src/types/sessionRecords';
+import type { possibleEntityNames } from 'src/types/session-records';
 import Avatar from 'antd/lib/avatar/avatar';
-import RelativeTo, { HumanizeDuration } from './Datetime/relativeTime';
+import RelativeTo, { HumanizeDuration } from './Datetime/relative-time';
 
-export default function RenderTimeRelativeToStart(props: {
+export default function RenderTimeRelativeToStart(properties: {
     value?: [Dayjs, Dayjs];
     style?: CSSProperties;
 }): ReactNode {
-    if (props.value == null) return <></>;
+    if (properties.value == undefined) return <></>;
     return (
         <RelativeTo
-            dateTime={props.value[0]}
-            wrt={props.value[1]}
-            style={props.style}
+            dateTime={properties.value[0]}
+            wrt={properties.value[1]}
+            style={properties.style}
         />
     );
 }
 
-export function RenderDuration(props: {
+export function RenderDuration(properties: {
     value: Duration;
     style?: CSSProperties;
 }): ReactNode {
-    return <HumanizeDuration duration={props.value} style={props.style} />;
+    return (
+        <HumanizeDuration
+            duration={properties.value}
+            style={properties.style}
+        />
+    );
 }
 
-export function RenderStatus(props: {
+export function RenderStatus(properties: {
     value: string;
     marginTop?: string;
 }): ReactNode {
-    switch (props.value as statusOfEntity) {
+    switch (properties.value as statusOfEntity) {
         case 'PASSED': {
             return (
                 <CheckCircleFilled
                     style={{
                         fontSize: '16px',
                         color: 'green',
-                        marginTop: props.marginTop,
+                        marginTop: properties.marginTop,
                     }}
                     title="Passed"
                 />
@@ -56,7 +61,7 @@ export function RenderStatus(props: {
                     style={{
                         fontSize: '16px',
                         color: 'red',
-                        marginTop: props.marginTop,
+                        marginTop: properties.marginTop,
                     }}
                     title="Failed"
                 />
@@ -68,7 +73,7 @@ export function RenderStatus(props: {
                     style={{
                         fontSize: '16px',
                         color: 'yellow',
-                        marginTop: props.marginTop,
+                        marginTop: properties.marginTop,
                     }}
                     title="Skipped"
                 />
@@ -80,7 +85,7 @@ export function RenderStatus(props: {
                     style={{
                         fontSize: '16px',
                         color: 'yellow',
-                        marginTop: props.marginTop,
+                        marginTop: properties.marginTop,
                     }}
                     title="Pending"
                 />
@@ -92,7 +97,7 @@ export function RenderStatus(props: {
                     style={{
                         fontSize: '16px',
                         color: 'orangered',
-                        marginTop: props.marginTop,
+                        marginTop: properties.marginTop,
                     }}
                     spin
                     title="Retried Suite"
@@ -102,8 +107,10 @@ export function RenderStatus(props: {
     }
 }
 
-export function RenderEntityType(props: { entityName: string }): ReactNode {
-    switch (props.entityName.toLowerCase() as possibleEntityNames) {
+export function RenderEntityType(properties: {
+    entityName: string;
+}): ReactNode {
+    switch (properties.entityName.toLowerCase() as possibleEntityNames) {
         case 'chrome': {
             return <Avatar src={'/chrome.png'} size="small" />;
         }
@@ -114,13 +121,15 @@ export function RenderEntityType(props: { entityName: string }): ReactNode {
             return <Avatar src={'/edge.png'} size="small" />;
         }
         default: {
-            return <>{props.entityName?.toLocaleUpperCase()}</>;
+            return <>{properties.entityName?.toLocaleUpperCase()}</>;
         }
     }
 }
 
-export function RenderSystemType(props: { systemName: string }): ReactNode {
-    const target = props.systemName.toLowerCase();
+export function RenderSystemType(properties: {
+    systemName: string;
+}): ReactNode {
+    const target = properties.systemName.toLowerCase();
 
     if (target.startsWith('win'))
         return <Avatar src={'/windows.png'} size="large" />;