+++ graspit-reporters/packages/graspit-dashboard/CHANGELOG.md
@@ -1,5 +1,23 @@
 # graspit
 
+## 3.4.1
+
+### Patch Changes
+
+- build before publish
+
+## 3.4.0
+
+### Minor Changes
+
+- refractor: eslint fixes
+
+## 3.3.1
+
+### Patch Changes
+
+- fix type issues in gantt charts
+
 ## 3.3.0
 
 ### Minor Changes